⭐️ allow and deny subscriptions + scan VMs options
This PR adds three more flags to the experimental command
`cnspec integrate azure`.

`--allow`, `--deny` used to control the integration of subscriptions.

`--scan-vms` that enables scanning VMs on Azure subscriptions via
RunCommand.

Additionally, I moved the `onboarding` code that is generic to its own
package so that new integration commands like `gcp` or `aws` can
leverage it.
