+++ internal/onboarding/constants.go
@@ -1,0 +1,12 @@
+// Copyright (c) Mondoo, Inc.
+// SPDX-License-Identifier: BUSL-1.1
+
+package onboarding
+
+const (
+	mondooProviderSource  = "mondoohq/mondoo"
+	mondooProviderVersion = "~> 0.19"
+
+	requiredTerraformVersion = ">= 1.0.11"
+	installTerraformVersion  = "1.9.8"
+)
